requirements.txt
Django==5.2.1
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.8
supabase==2.15.2
python-jose[cryptography]==3.5.0
requests==2.32.3
gunicorn==21.2.0
whitenoise==6.5.0
openai>=1.40.0,<2.0.0
azure-search-documents==11.5.2
python-dotenv
langchain>=0.3.15
langchain-openai>=0.2.12
langchain-core>=0.3.31
langgraph>=0.2.58
